+++ Projects/Team1/TestFrameworkForOpenCart/TestFramework.OpenCart/TestFramework.OpenCart/Data/DBTest.cs
@@ -11,10 +11,27 @@ namespace TestFramework.OpenCart
 {
     class DBTest
     {
-        //private static readonly string _connectionString = @"Data Source=127.0.0.1;port=3309;username=admin;password=admin;database=storedb;";
         private static readonly string _connectionString = @"Data Source=127.0.0.1;port=3309;username=admin;password=admin;Initial Catalog=storedb;Integrated Security=True";
-
-        public static string ExecuteQuery(string queryString)
+        public static void ExecuteQuery(string queryString)
+        {
+            try
+            {
+                using (var databaseConnection = new MySqlConnection(_connectionString))
+                {
+                    databaseConnection.Open();
+                    using (var command = new MySqlCommand(queryString, databaseConnection))
+                    {
+                        command.ExecuteScalar();
+                    }
+                }
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.Message);
+                //throw ex;
+            }
+        }
+        public static string ExecuteSelectQuery(string queryString)
         {
             try
             {
@@ -34,7 +51,6 @@ namespace TestFramework.OpenCart
             }
             catch (Exception ex)
             {
-                // Show any error message.
                 Console.WriteLine(ex.Message);
                 throw ex;
             }
@@ -42,36 +58,41 @@ namespace TestFramework.OpenCart
         public static string CheckSession(string key)
         {
             var queryString = $@"SELECT `date_added` FROM `oc_api_session` WHERE `session_id` = '{key}'";
-            return ExecuteQuery(queryString);
+            return ExecuteSelectQuery(queryString);
         }
         public static string CheckAddingToCart(string key)
         {
             var queryString = $@"SELECT `product_id`  FROM `oc_cart` WHERE `session_id` = '{key}'";
             var secondqueryString = $@"SELECT `quantity`  FROM `oc_cart` WHERE `session_id` = '{key}'";
-            return  ExecuteQuery(queryString) + ExecuteQuery(secondqueryString);
+            return ExecuteSelectQuery(queryString) + ExecuteSelectQuery(secondqueryString);
         }
         public static string CheckCartOutput(string key)
         {
             var queryString = $@"SELECT `product_id`  FROM `oc_cart` WHERE `session_id` = '{key}'";
-            return ExecuteQuery(queryString);
+            return ExecuteSelectQuery(queryString);
         }
         public static string CheckCartChange(string key)
         {
             var queryString = $@"SELECT `quantity`  FROM `oc_cart` WHERE `session_id` = '{key}'";
-            return ExecuteQuery(queryString);
+            return ExecuteSelectQuery(queryString);
         }
         public static string CheckCartEmpty(string key)
         {
             try 
             {
                 var queryString = $@"SELECT *  FROM `oc_cart` WHERE `session_id` = '{key}'";
-                ExecuteQuery(queryString);
+                ExecuteSelectQuery(queryString);
             }
             catch(NullReferenceException)
             {
                 return "Empty";
             }
             return "";
+        }
+        public static void DeleteCustomer(string firstName, string lastName, string email, string telephone)
+        {
+            var queryString = $@"DELETE FROM `oc_customer` WHERE `firstname` = '{firstName}' AND `lastname` = '{lastName}' AND `email` = '{email}' AND `telephone` = '{telephone}'";
+            ExecuteQuery(queryString);
         }
     }
 }
